assert syncside params in proper order
Without going into too much headachy calculus, the order in which the
r3cell parameters are passed in apparently matters for my algorithm.
This, in part, is coupled to how I order the oanchors, in particular
'rotateblock'. Instead of abelian-ifying my algorithm (which I'm not
sure if there's an easy way to do, other than altering the inputs
pre-processing), I am making assertions that they are coming in with the
proper order. This has the added benefit that, at compile time, I should
be able to catch instances of where I break my own rule, thus speeding
up execution of the module once perfected.

The order in which the inputs are brought in matters as so: the first
r3cell must be to the left of the 2nd r3cell, when oriented in such a
way that the vector of advance will be downward. Thankfully, it's easy
enough to check this in the counter/clockwise blocks, since it's an
equivalent criteria.