terminate flow on failed assertions
This is needed in order to convince the compiler that we won't be using
variables which haven't been initialized. Assertions take care of this,
normally, but non-assert builds will compile in error otherwise.

Consider rewriting some of the static functions to return numeric values
indicating its success, and then checking these values in the caller.
Is there enough risk to warrant the overhead? Vs. compiling with
assertions to ensure codebase accuracy.