implement row,col translation feature
This is necessary due to how I represent the <row,col> orientation
per-side, in particular on the back-side of the cube. For example, the
front-facing side's first column, is actually the back-facing side's
last-column. Thus, for the algorithm to work appropriately, I have to be
able to translate these indices.

This isn't yet fully tested; I could have missed something.