allow builds without TAP support
This commit allows r3 to automatically build with or without libcheck TAP
functionality, depending on the version of libcheck.

Historically, r3 has unconditionally used TAP for its build tests, but this
limited `make check` functionality from running on platforms without newer
versions of libcheck. This patch will, at configure time, determine if the
available version of libcheck is capable of supporting TAP, and if not,
compile-out the TAP functionality.