remove r3_cube_get_face
This commit removes the r3_cube_get_face API call. At the current
maturity of the baseline, it is questionable as to whether or not this
function call should be supported. Depending on the direction r3 is
carried (e.g., whether a side should have a set of API calls that
operate on it, possibly include manipulating a user-supplied ptr/value),
this call may be re-introduced.