remove use of s1ct * sct for next-cell calculation
As the body of this patch suggests, this is an experimental addition to
the codebase. It is somewhat surprising that unit tests continue to pass
with the s1ct * sct translation omitted.

If future unit tests (that better check for expected state) tend to fail
without apparent reason, consider reverting this patch to see if that
helps in any way.