admit defeat on ck_assert_int_eq for backwards compat
It'd be nice to use ck_assert_int_eq, since, you know, it's available as
part of the check API and gives better insight into what's going on, but
evidently some older versions of libcheck are susceptible to a bug that
causes args to be evalulated more than once. This disrupts state and
causes wonkiness when arguments to the macro have stateful side-effects.

It's very tempting to just bump the required version of check -- and I
may very well end up doing that in the long run -- but for now, I have
reason to support legacy versions of check.

See the check repo, commit 10064e56e3fb9099d347da4f2ef8d3e6375cf994
("Fixes 3035490."), for more info about the specific defect this is
mitigating.

    <https://github.com/libcheck/check.git>